requirements.txt
boto3==1.9.*
botocore==1.12.*
celery==4.2.*
dj-database-url==0.5.*
django==2.0.*
django-auditlog==0.4.*
django-cors-headers==2.4.*
django-choices==1.6.*
django-environ==0.4.*
django_extensions==2.1.*
djangorestframework==3.9.*
drf-nested-routers==0.91
drf_yasg==1.11.*
dry_rest_permissions==0.1.*
jinja2==2.10
python-dateutil==2.6.0
python-magic==0.4.*
requests==2.21.*
whitenoise==3.3.*
psycopg2-binary==2.7.*
pycpfcnpj
coverage==4.5.*
flake8==3.7.*
pycodestyle==2.5.*
pyflakes==2.1.*
pytest==4.3.*
pytest-cov==2.6.*
pytest-django==3.4.*
pytest-mock==1.10.*
requests-mock==1.5.*
